requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
requests==2.31.0
six==1.16.0
uplink==0.9.7
uritemplate==4.1.1
urllib3==2.0.2
async-timeout==4.0.2
redis==4.5.5
spotipy==2.23.0
python-dotenv==1.0.0